Доделан поиск, переделан просмотр продуктов

+++ bot/handlers/market.py
@@ -1,22 +1,28 @@
 from asyncio import sleep
+from random import choice
 
-from telebot.types import CallbackQuery, Message
+from telebot.types import CallbackQuery, Message, InlineKeyboardMarkup
 
 from bot.config import mongo_client
 from bot.exec import bot
-from bot.modules.data_format import (list_to_inline, list_to_keyboard,
-                                     seconds_to_str, user_name, escape_markdown)
+from bot.modules.data_format import (escape_markdown, list_to_inline,
+                                     list_to_keyboard, seconds_to_str,
+                                     user_name)
 from bot.modules.item import AddItemToUser, counts_items, get_name, item_info
 from bot.modules.localization import get_data, get_lang, t
-from bot.modules.market import (add_product, create_seller, delete_product,
-                                buy_product, product_ui, seller_ui)
-from bot.modules.market_chose import (pr_edit_description, pr_edit_name,
-                                      prepare_add, prepare_data_option,
-                                      prepare_delete_all, prepare_edit_price, pr_edit_image, buy_item, promotion_prepare)
+from bot.modules.market import (add_product, buy_product, create_seller,
+                                delete_product, preview_product, product_ui,
+                                seller_ui, create_push)
+from bot.modules.market_chose import (buy_item, find_prepare,
+                                      pr_edit_description, pr_edit_image,
+                                      pr_edit_name, prepare_add,
+                                      prepare_data_option, prepare_delete_all,
+                                      prepare_edit_price, promotion_prepare,
+                                      send_info_pr)
 from bot.modules.markup import answer_markup, cancel_markup, count_markup
 from bot.modules.markup import markups_menu as m
-from bot.modules.states_tools import (ChooseOptionState, ChooseStepState,
-                                      prepare_steps)
+from bot.modules.states_tools import (ChooseOptionState, ChoosePagesState,
+                                      ChooseStepState, prepare_steps)
 from bot.modules.user import premium
 
 users = mongo_client.bot.users
@@ -25,6 +31,7 @@ tavern = mongo_client.connections.tavern
 sellers = mongo_client.market.sellers
 items = mongo_client.bot.items
 products = mongo_client.market.products
+puhs = mongo_client.market.puhs
 
 async def create_adapter(return_data, transmitted_data):
     chatid = transmitted_data['chatid']
@@ -94,10 +101,10 @@ async def add_product_com(message: Message):
     chatid = message.chat.id
 
     options = {
-        "🍕 ➡ 🪙": 'items_coins',
-        "🪙 ➡ 🍕": 'coins_items',
-        "🍕 ➡ 🍕": 'items_items',
-        "🍕 ➡ ⏳": 'auction'
+        "🍕 ➞ 🪙": 'items_coins',
+        "🪙 ➞ 🍕": 'coins_items',
+        "🍕 ➞ 🍕": 'items_items',
+        "🍕 ➞ ⏳": 'auction'
     }
 
     b_list = list(options.keys())
@@ -114,18 +121,24 @@ async def my_products(message: Message):
     lang = get_lang(message.from_user.id)
     chatid = message.chat.id
 
-    user_prd = list(products.find({'owner_id': userid}, {'_id': 1}))
-    if user_prd:
-        for i in user_prd:
+    user_prd = list(products.find({'owner_id': userid}))
+    rand_p = {}
 
-            text, markup = product_ui(lang, i['_id'], True)
-            await bot.send_message(chatid, text, reply_markup=markup, parse_mode='Markdown')
-            await sleep(1)
+    if user_prd:
+        for product in user_prd:
+            rand_p[
+                preview_product(product['items'], product['price'], 
+                                product['type'], lang)
+            ] = product['_id']
+
+        await bot.send_message(chatid, t('products.search', lang))
+        await ChoosePagesState(send_info_pr, userid, chatid, lang, rand_p, 1, 3, 
+                               None, False, False)
     else:
         text = t('no_products', lang)
         await bot.send_message(chatid, text,  parse_mode='Markdown')
 
-@bot.callback_query_handler(func=lambda call: call.data.startswith('product_info'), private=True)
+@bot.callback_query_handler(func=lambda call: call.data.startswith('product_info'))
 async def product_info(call: CallbackQuery):
     call_data = call.data.split()
     chatid = call.message.chat.id
@@ -134,39 +147,49 @@ async def product_info(call: CallbackQuery):
 
     call_type = call_data[1]
     alt_id = call_data[2]
+    product = products.find_one({'alt_id': alt_id})
+    if product:
+        if call_type == 'delete':
+            if product['owner_id'] == userid:
+                status = await delete_product(None, alt_id)
 
-    if call_type == 'delete':
-        status = await delete_product(None, alt_id)
+                if status: text = t('product_info.delete', lang)
+                else: text = t('product_info.error', lang)
 
-        if status: text = t('product_info.delete', lang)
-        else: text = t('product_info.error', lang)
-
-        markup = list_to_inline([])
-        await bot.edit_message_text(text, chatid, call.message.id, reply_markup=markup, parse_mode='Markdown')
-    else:
-        product = products.find_one({'alt_id': alt_id})
-        if product:
-            if call_type == 'edit_price':
+                markup = list_to_inline([])
+                await bot.edit_message_text(text, chatid, call.message.id, reply_markup=markup, parse_mode='Markdown')
+        else:
+            if call_type == 'edit_price' and product['owner_id'] == userid:
                 await prepare_edit_price(userid, chatid, lang, alt_id)
 
-            elif call_type == 'add':
+            elif call_type == 'add' and product['owner_id'] == userid:
                 await prepare_add(userid, chatid, lang, alt_id)
 
             elif call_type == 'items':
+                itm = []
                 for item in product['items']:
-                    text, image = item_info(item, lang)
-                    await bot.send_message(chatid, text, parse_mode='Markdown')
+                    if item not in itm:
+                        itm.append(item)
+                        text, image = item_info(item, lang)
 
-            elif call_type == 'buy':
-                await buy_item(userid, chatid, lang, product, user_name(call.from_user))
+                        try:
+                            await bot.send_photo(chatid, image, text, 'Markdown')
+                        except:
+                            await bot.send_message(chatid, text, parse_mode='Markdown')
+
+            elif call_type == 'buy' and product['owner_id'] != userid:
+                if product['owner_id'] != userid:
+                    await buy_item(userid, chatid, lang, product, 
+                                   user_name(call.from_user), call.message.id)
 
             elif call_type == 'info':
                 text, markup = product_ui(lang, product['_id'], 
                                           product['owner_id'] == userid)
-                await bot.send_message(chatid, text, reply_markup=markup, parse_mode='Markdown')
+                await bot.send_message(userid, text, reply_markup=markup, parse_mode='Markdown')
 
-            elif call_type == 'promotion':
-                await promotion_prepare(userid, chatid, lang, product['_id'], call.message.id)
+            elif call_type == 'promotion' and product['owner_id'] == userid:
+                await promotion_prepare(userid, chatid, lang, product['_id'], 
+                                        call.message.id)
 
 @bot.callback_query_handler(func=lambda call: call.data.startswith('seller'), private=True)
 async def seller(call: CallbackQuery):
@@ -176,7 +199,7 @@ async def seller(call: CallbackQuery):
     lang = get_lang(call.from_user.id)
 
     call_type = call_data[1]
-    owner_id = call_data[2]
+    owner_id = int(call_data[2])
 
     # Кнопки вызываемые владельцем
     if call_type == 'cancel_all':
@@ -193,14 +216,89 @@ async def seller(call: CallbackQuery):
 
     # Кнопки вызываемые не владельцем
     elif call_type == 'info':
-        my_status = int(owner_id) == userid
-        text, markup, image = seller_ui(userid, lang, False)
-        await bot.send_photo(chatid, image, text, parse_mode='Markdown', reply_markup=markup)
+        my_status = owner_id == userid
+
+        seller = sellers.find_one({'owner_id': owner_id})
+        if seller:
+            try:
+                chat_user = await bot.get_chat_member(seller['owner_id'], 
+                                                      seller['owner_id'])
+                name = user_name(chat_user.user)
+            except: name = '-'
+
+            text, markup, image = seller_ui(owner_id, lang, my_status, name)
+            await bot.send_photo(chatid, image, text, parse_mode='Markdown', reply_markup=markup)
 
     elif call_type == 'all':
-        user_prd = list(products.find({'owner_id': int(owner_id)}, {'_id': 1}))
-        if user_prd:
-            for i in user_prd:
-                text, markup = product_ui(lang, i['_id'], False)
-                await bot.send_message(chatid, text, reply_markup=markup, parse_mode='Markdown')
-                await sleep(1)
+        user_prd = list(products.find({'owner_id': owner_id}))
+
+        rand_p = {}
+        for product in user_prd:
+            rand_p[
+                preview_product(product['items'], product['price'], 
+                                product['type'], lang)
+            ] = product['_id']
+
+        await bot.send_message(chatid, t('products.search', lang))
+        await ChoosePagesState(send_info_pr, userid, chatid, lang, rand_p, 1, 3, 
+                               None, False, False)
+
+@bot.message_handler(text='commands_name.market.random', is_authorized=True)
+async def random_products(message: Message):
+    userid = message.from_user.id
+    lang = get_lang(message.from_user.id)
+    chatid = message.chat.id
+
+    products_all = list(products.find({"owner_id": {"$ne": userid}})).copy()
+    rand_p = {}
+
+    if products_all:
+        for _ in range(18):
+            if products_all:
+                prd = choice(products_all)
+                products_all.remove(prd)
+
+                product = products.find_one({'_id': prd['_id']})
+                if product:
+                    rand_p[
+                        preview_product(product['items'], product['price'], 
+                                        product['type'], lang)
+                    ] = product['_id']
+            else: break
+
+        await bot.send_message(chatid, t('products.search', lang))
+        await ChoosePagesState(send_info_pr, userid, chatid, lang, rand_p, 1, 3, 
+                               None, False, False)
+    else:
+        await bot.send_message(chatid, t('products.null', lang))
+
+@bot.message_handler(text='commands_name.market.find', is_authorized=True)
+async def find_products(message: Message):
+    userid = message.from_user.id
+    lang = get_lang(message.from_user.id)
+    chatid = message.chat.id
+
+    await find_prepare(userid, chatid, lang)
+
+@bot.callback_query_handler(func=lambda call: call.data.startswith('create_push'), private=True)
+async def push(call: CallbackQuery):
+    call_data = call.data.split()
+    chatid = call.message.chat.id
+    userid = call.from_user.id
+    lang = get_lang(call.from_user.id)
+
+    channel_id = int(call_data[1])
+
+    res = puhs.find_one({'owner_id': userid})
+    if res:
+        puhs.update_one({'owner_id': userid}, 
+                        {'channel_id': channel_id, 'lang': lang})
+        text = t('push.update', lang)
+    else: 
+        create_push(userid, channel_id, lang)
+        text = t('push.new', lang)
+
+    await bot.send_message(userid, text)
+    await bot.edit_message_reply_markup(chatid, call.message.id, 
+                                        reply_markup=InlineKeyboardMarkup())
+
